pyproject.toml
[tool.poetry]
name = "proxy-scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
httpx = {extras = ["socks"], version = "0.26.0"}
beautifulsoup4 = "^4.9.3"
socksio = "^1.0.0"
python-dotenv = "^0.15.0"
"flask-caching" = "^2.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
